filtro

+++ app/src/main/java/fiap/com/br/fiapapp/view/Cadastro.kt
@@ -1,14 +1,10 @@
 package fiap.com.br.fiapapp.view
 
-import android.app.Activity
 import android.content.DialogInterface
-import android.content.Intent
 import android.icu.text.SimpleDateFormat
 import android.os.Build
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.provider.ContactsContract
-import android.util.Log
 import android.view.View
 import android.widget.*
 import androidx.annotation.RequiresApi
@@ -17,6 +13,7 @@ import androidx.core.graphics.drawable.toDrawable
 import fiap.com.br.fiapapp.R
 import fiap.com.br.fiapapp.model.Veiculo
 import fiap.com.br.fiapapp.presenter.*
+import fiap.com.br.fiapapp.presenter.interfaces.*
 import java.util.*
 import kotlin.collections.ArrayList
 
@@ -146,7 +143,7 @@ ModeloContrato.ListaModeloView, CorContrato.ListaCorView, FilialContrato.ListaFi
                 presenterCor.obterCorSelecionada(nomeCor)
 
                 presenterVeiculo.cadastrarVeiculo(
-                    codModelo, codCor, cnpjEmpresa.toLong(), km.toLong(),
+                    codModelo, codCor, cnpjEmpresa.toLong(), km.toInt(),
                     valor.toDouble(), placa, detalhes
                 )
             }
